+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * New cop ClassMethods checks for uses for class/module names in definitions of class/module methods
 * New cop SingleLineMethods checks for methods implemented on a single line
 * New cop FavorJoin checks for usages of Array#* with a string argument
+* New cop BlockComments checks tracks uses of block comments(=begin/=end comments)
 
 ### Bugs fixed